Move formatTimestamp.ts into utility folder
